refactor(MessagesUiState): Message ui 상태를 변경하는 로직을 MessagesUiState로 옮김
